move prepare_* out of main model
